LPD-56913 Rename and fix tests
