+++ modules/apps/site/site-cms-site-initializer-test/src/testIntegration/java/com/liferay/site/cms/site/initializer/internal/display/context/test/ViewContentsSectionDisplayContextTest.java
@@ -7,6 +7,7 @@ package com.liferay.site.cms.site.initializer.internal.display.context.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.fragment.renderer.FragmentRenderer;
+import com.liferay.frontend.data.set.model.FDSActionDropdownItem;
 import com.liferay.object.constants.ObjectDefinitionConstants;
 import com.liferay.object.constants.ObjectEntryFolderConstants;
 import com.liferay.object.constants.ObjectFolderConstants;
@@ -28,6 +29,7 @@ import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
 import jakarta.servlet.http.HttpServletRequest;
 
+import java.util.List;
 import java.util.Map;
 
 import org.junit.Assert;
@@ -40,12 +42,12 @@ import org.junit.runner.RunWith;
 import org.springframework.mock.web.MockHttpServletResponse;
 
 /**
- * @author Eudaldo Alonso
+ * @author Mikel Lorza
  */
 @FeatureFlag("LPD-17564")
 @RunWith(Arquillian.class)
 @Sync
-public class FilesSectionDisplayContextTest
+public class ViewContentsSectionDisplayContextTest
 	extends BaseSectionDisplayContextTestCase {
 
 	@ClassRule
@@ -62,22 +64,24 @@ public class FilesSectionDisplayContextTest
 		Map<String, String> expectedResultMap = LinkedHashMapBuilder.put(
 			"folder", StringPool.BLANK
 		).put(
-			"Basic Document",
+			"Basic Web Content",
 			getHref(
 				objectDefinitionLocalService.
 					fetchObjectDefinitionByExternalReferenceCode(
-						"L_BASIC_DOCUMENT", TestPropsValues.getCompanyId()))
+						"L_BASIC_WEB_CONTENT", TestPropsValues.getCompanyId()))
 		).build();
 
 		testGetCreationMenu(
 			ReflectionTestUtil.invoke(
-				_getFilesSectionDisplayContext(getMockHttpServletRequest()),
+				_getViewContentsSectionDisplayContext(
+					getMockHttpServletRequest()),
 				"getCreationMenu", new Class<?>[0]),
 			expectedResultMap);
 
 		ObjectFolder objectFolder =
 			objectFolderLocalService.fetchObjectFolderByExternalReferenceCode(
-				ObjectFolderConstants.EXTERNAL_REFERENCE_CODE_FILE_TYPES,
+				ObjectFolderConstants.
+					EXTERNAL_REFERENCE_CODE_CONTENT_STRUCTURES,
 				TestPropsValues.getCompanyId());
 
 		ObjectDefinition objectDefinition = addCustomObjectDefinition(
@@ -112,29 +116,81 @@ public class FilesSectionDisplayContextTest
 
 		testGetCreationMenu(
 			ReflectionTestUtil.invoke(
-				_getFilesSectionDisplayContext(getMockHttpServletRequest()),
+				_getViewContentsSectionDisplayContext(
+					getMockHttpServletRequest()),
 				"getCreationMenu", new Class<?>[0]),
 			expectedResultMap);
 	}
 
-	private Object _getFilesSectionDisplayContext(
+	@Test
+	public void testGetFDSActionDropdownItems() throws Exception {
+		List<FDSActionDropdownItem> fdsActionDropdownItems =
+			_getFDSActionDropdownItems();
+
+		Assert.assertEquals(
+			fdsActionDropdownItems.toString(), 5,
+			fdsActionDropdownItems.size());
+
+		_assertFDSActionDropdownItem(
+			fdsActionDropdownItems.get(0), "view", "actionLinkFolder",
+			"view-folder", "get", "item");
+		_assertFDSActionDropdownItem(
+			fdsActionDropdownItems.get(1), "pencil", "editFolder", "edit",
+			"get", "item");
+		_assertFDSActionDropdownItem(
+			fdsActionDropdownItems.get(2), "pencil", "actionLink", "edit",
+			"get", "item");
+		_assertFDSActionDropdownItem(
+			fdsActionDropdownItems.get(3), "password-policies", "permissions",
+			"permissions", "get", "item");
+		_assertFDSActionDropdownItem(
+			fdsActionDropdownItems.get(4), "trash", "delete", "delete",
+			"delete", "item");
+	}
+
+	private void _assertFDSActionDropdownItem(
+		FDSActionDropdownItem fdsActionDropdownItem, String icon, String id,
+		String label, String method, String type) {
+
+		Assert.assertNotNull(fdsActionDropdownItem);
+
+		Map<String, String> data =
+			(Map<String, String>)fdsActionDropdownItem.get("data");
+
+		Assert.assertEquals(id, data.get("id"));
+		Assert.assertEquals(method, data.get("method"));
+
+		Assert.assertEquals(icon, fdsActionDropdownItem.get("icon"));
+		Assert.assertEquals(label, fdsActionDropdownItem.get("label"));
+		Assert.assertEquals(type, fdsActionDropdownItem.get("type"));
+	}
+
+	private List<FDSActionDropdownItem> _getFDSActionDropdownItems()
+		throws Exception {
+
+		return ReflectionTestUtil.invoke(
+			_getViewContentsSectionDisplayContext(getMockHttpServletRequest()),
+			"getFDSActionDropdownItems", new Class<?>[0]);
+	}
+
+	private Object _getViewContentsSectionDisplayContext(
 			HttpServletRequest httpServletRequest)
 		throws Exception {
 
 		_fragmentRenderer.render(
 			null, httpServletRequest, new MockHttpServletResponse());
 
-		Object filesSectionDisplayContext = httpServletRequest.getAttribute(
+		Object contentsSectionDisplayContext = httpServletRequest.getAttribute(
 			"com.liferay.site.cms.site.initializer.internal.display.context." +
-				"FilesSectionDisplayContext");
+				"ViewContentsSectionDisplayContext");
 
-		Assert.assertNotNull(filesSectionDisplayContext);
+		Assert.assertNotNull(contentsSectionDisplayContext);
 
-		return filesSectionDisplayContext;
+		return contentsSectionDisplayContext;
 	}
 
 	@Inject(
-		filter = "component.name=com.liferay.site.cms.site.initializer.internal.fragment.renderer.FilesSectionFragmentRenderer"
+		filter = "component.name=com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewContentsSectionFragmentRenderer"
 	)
 	private FragmentRenderer _fragmentRenderer;
